CreatedInterfaceStructureForSampleProgram
Purpose is to later while we progress in the project and we can add new similarity methods defined in the interface and we can change similarity method and Visualisation method if we want by adding implementation in service
