+++ SemanticaAnalysisTextualData/SemanticaAnalysisTextualData/SemanticaAnalysisTextualData.csproj
@@ -14,19 +14,18 @@
 
   <ItemGroup>
     <PackageReference Include="App.Metrics.Health.Extensions.Configuration" Version="3.2.0" />
+    <PackageReference Include="Azure.AI.OpenAI" Version="2.1.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="9.0.0" />
 	  <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="OpenAI" Version="2.1.0" />
     <PackageReference Include="OpenAI-DotNet" Version="8.4.1" />
+    <PackageReference Include="System.Net.Http.Json" Version="9.0.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Documentation\" />
     <Folder Include="Images\" />
-    <Folder Include="docs\" />
-    <Folder Include="data\" />
     <Folder Include="ErrorsWarningFacedScreenshots\" />
-    <Folder Include="Results\" />
   </ItemGroup>
 
 </Project>
